PLANNER-2212 Minimize the occurrence of match weight exceptions (#982)
